clean up test bins, improve dns tests
